Add new tools, refactor so can use on mac

+++ timelapse.py
@@ -60,7 +60,7 @@ class TimelapseConfig(Config):
                                 'FORMAT: comma separated integers between 0 and 255, no spaces EX: "R,G,B" ')
 
         timelapse = parser.add_argument_group('Timelapse', 'Parameters that control timelapse')
-        timelapse.add_argument('--framerate', type=int, default=None,
+        timelapse.add_argument('--framerate', type=float, default=None,
                                help='When "all-frames" is set, "framerate" limits how often a new frame is taken. '
                                     'Int value. Units is seconds. EX. Setting framerate to "3" will take a frame every'
                                     '3 seconds. Defaults to 0 which means "as fast as you can" ')
